+++ TresurHuntFrontend/app/build.gradle
@@ -42,5 +42,5 @@ dependencies {
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     implementation 'com.android.volley:volley:1.2.0'
     implementation 'com.github.bumptech.glide:glide:4.12.0'
-
+    implementation 'pl.droidsonroids.gif:android-gif-drawable:1.2.19'
 }
