+++ TresurHuntFrontend/app/src/main/java/com/example/myapplication/models/GlobalClass.java
@@ -1,4 +1,4 @@
-package com.example.myapplication;
+package com.example.myapplication.models;
 
 import android.app.Application;
